Complexity sliders
Changed "sparseness" to a more understandable "complexity." The higher
the value, the more dense with notes the melody will be. Also added a
pair of sliders for the minimum and maximum value that complexity can
have. If no available rhythms are within the range, it will choose a
random rhythm.

Also added more rhythms to the pool of pre-generated melodies, though
some are not so great. Still working on how to generate rhythms
procedurally.