Procedural rhythms
Very shaky start, but procedural rhythms are now implemented. This will
probably take a few rewrites before it sounds even halfway decent, but
it's a start. Furthermore, some big GUI changes in this version. I've
added a syncopation slider, but it's not having as strong an effect as
I'd like.